+++ src/main/java/com/gerenciador/gerenciadorDeTarefas/services/TarefaService.java
@@ -4,40 +4,57 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.gerenciador.gerenciadorDeTarefas.entities.Tarefa;
 import com.gerenciador.gerenciadorDeTarefas.repositories.TarefaRepository;
+import com.gerenciador.gerenciadorDeTarefas.services.exceptions.DatabaseException;
+import com.gerenciador.gerenciadorDeTarefas.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class TarefaService {
 
 	@Autowired
 	private TarefaRepository repositorio;
-	
+
 	public List<Tarefa> listarTarefas() {
-		return repositorio.findAll();	
+		return repositorio.findAll();
 	}
-	
+
 	public Tarefa buscarPorId(Long id) {
 		Optional<Tarefa> proj = repositorio.findById(id);
-		return proj.get();
+		return proj.orElseThrow(() -> new ResourceNotFoundException("Erro ao buscar tarefa."));
 	}
-	
+
 	public Tarefa salvarTarefa(Tarefa obj) {
 		return repositorio.save(obj);
 	}
-	
+
 	public void deletarTarefa(Long id) {
-		repositorio.deleteById(id);
-		
+		try {
+			if(! repositorio.existsById(id)) throw new ResourceNotFoundException("Erro ao deletar tarefa.");
+			repositorio.deleteById(id);
+		} catch(EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(e.getMessage());
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+
 	}
-	
+
 	public Tarefa atualizarTarefa(Long id, Tarefa obj) {
-		Tarefa tarefa = repositorio.getReferenceById(id);
-		updateTarefa(tarefa, obj);
-		
-		return repositorio.save(obj);
+		try {
+			Tarefa tarefa = repositorio.getReferenceById(id);
+			updateTarefa(tarefa, obj);
+
+			return repositorio.save(obj);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(e.getMessage());
+		}
 	}
 
 	private void updateTarefa(Tarefa tarefa, Tarefa obj) {
