+++ src/main/java/com/gerenciador/gerenciadorDeTarefas/services/UsuarioService.java
@@ -4,10 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.gerenciador.gerenciadorDeTarefas.entities.Usuario;
 import com.gerenciador.gerenciadorDeTarefas.repositories.UsuarioRepository;
+import com.gerenciador.gerenciadorDeTarefas.services.exceptions.DatabaseException;
+import com.gerenciador.gerenciadorDeTarefas.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UsuarioService {
@@ -21,7 +27,7 @@ public class UsuarioService {
 	
 	public Usuario buscarPorId(Long id) {
 		Optional<Usuario> usuario = repositorio.findById(id);
-		return usuario.get();
+		return usuario.orElseThrow(() -> new ResourceNotFoundException("Erro ao buscar usuário."));
 	}
 	
 	public Usuario salvarUsuario(Usuario obj) {
@@ -29,15 +35,25 @@ public class UsuarioService {
 	}
 	
 	public void deletarUsuario(Long id) {
-		repositorio.deleteById(id);
-		
+		try {
+			if(!repositorio.existsById(id)) throw new ResourceNotFoundException("Erro ao deletar usuário.");
+			repositorio.deleteById(id);
+		} catch(EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(e.getMessage());
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public Usuario atualizarUsuario(Long id, Usuario obj) {
-		Usuario user = repositorio.getReferenceById(id);
-		updateUsuario(user, obj);
-		
-		return repositorio.save(obj);
+		try {
+			Usuario user = repositorio.getReferenceById(id);
+			updateUsuario(user, obj);
+			
+			return repositorio.save(obj);
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException(e.getMessage());
+		}
 	}
 
 	private void updateUsuario(Usuario user, Usuario obj) {
